requirements.txt
Django==2.1.9
pyTelegramBotAPI==3.6.6
django-jsonfield==1.3.1
